requirements.txt
flask
waitress
python-dotenv
mysql-connector-python
bcrypt
pyjwt
